app/backend/requirements.txt

flask
prometheus_client
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-exporter-console
opentelemetry-instrumentation-flask
